fix(rc.local.sh.tpl): adiciona verificação para o symlink /dev/kmsg e reinicia microk8s se necessário

+++ scripts/rc.local.sh.tpl
@@ -3,8 +3,13 @@
 cat > /etc/rc.local <<EOF
 #!/bin/bash
 
-ln -s /dev/console /dev/kmsg
-exit 0
+# If this symlink is missing, likely microk8s isn't running
+if [ ! -L /dev/kmsg ]; then
+  ln -s /dev/console /dev/kmsg
+  apparmor_parser -r /var/lib/snapd/apparmor/profiles/*
+  microk8s stop
+  microk8s start
+fi
 EOF
 
 chmod +x /etc/rc.local
